# openapi.yaml
openapi: 3.0.1
info:
  title: Tech Inventory API
  version: v1
paths:
...
  /api/v1/imports/preview:
    post:
      tags:
        - Imports
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                File:
                  type: string
                  format: binary
                Mapping:
                  type: string
            encoding:
              File:
                style: form
              Mapping:
                style: form
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PreviewImportResult'
            application/json:
              schema:
                $ref: '#/components/schemas/PreviewImportResult'
            text/json:
              schema:
                $ref: '#/components/schemas/PreviewImportResult'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
        '413':
          description: Content Too Large
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/imports/commit:
    post:
      tags:
        - Imports
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                File:
                  type: string
                  format: binary
                Mapping:
                  type: string
            encoding:
              File:
                style: form
              Mapping:
                style: form
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/CommitImportResult'
            application/json:
              schema:
                $ref: '#/components/schemas/CommitImportResult'
            text/json:
              schema:
                $ref: '#/components/schemas/CommitImportResult'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '413':
          description: Content Too Large
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflict
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/v1/imports:
    get:
      tags:
        - Imports
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
            default: 25
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ImportStatus'
        - name: createdAfter
          in: query
          schema:
            type: string
            format: date-time
        - name: createdBefore
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ImportBatchSummaryResponsePagedResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ImportBatchSummaryResponsePagedResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ImportBatchSummaryResponsePagedResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
  '/api/v1/imports/{id}':
    get:
      tags:
        - Imports
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ImportBatchDetailResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ImportBatchDetailResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ImportBatchDetailResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  schemas:
...
    CommitImportResult:
      type: object
      properties:
        batchId:
          type: string
          format: uuid
        totalRows:
          type: integer
          format: int32
        importedRows:
          type: integer
          format: int32
        invalidRows:
          type: integer
          format: int32
        failedRows:
          type: array
          items:
            $ref: '#/components/schemas/ImportRowError'
          nullable: true
      additionalProperties: false
...
    ImportBatchDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
          nullable: true
        importedBy:
          type: string
          nullable: true
        rowCount:
          type: integer
          format: int32
        successCount:
          type: integer
          format: int32
        errorCount:
          type: integer
          format: int32
        status:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ImportRowError'
          nullable: true
        isErrorLogTruncated:
          type: boolean
        omittedErrorCount:
          type: integer
          format: int32
        errorLog:
          type: string
          nullable: true
      additionalProperties: false
    ImportBatchSummaryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
          nullable: true
        importedBy:
          type: string
          nullable: true
        rowCount:
          type: integer
          format: int32
        successCount:
          type: integer
          format: int32
        errorCount:
          type: integer
          format: int32
        status:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      additionalProperties: false
    ImportBatchSummaryResponsePagedResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ImportBatchSummaryResponse'
          nullable: true
        totalCount:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
      additionalProperties: false
    ImportDevicePreview:
      type: object
      properties:
        name:
          type: string
          nullable: true
        brand:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        owner:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        serialNumber:
          type: string
          nullable: true
        network:
          type: string
          nullable: true
        purchaseDate:
          type: string
          format: date
          nullable: true
        purchasePrice:
          type: number
          format: double
          nullable: true
        currencyCode:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        retiredDate:
          type: string
          format: date
          nullable: true
        disposalMethod:
          type: string
          nullable: true
        purpose:
          type: string
          nullable: true
        operatingSystem:
          type: string
          nullable: true
        ipAddress:
          type: string
          nullable: true
        macAddress:
          type: string
          nullable: true
        productUrl:
          type: string
          nullable: true
        version:
          type: string
          nullable: true
      additionalProperties: false
    ImportFieldError:
      type: object
      properties:
        field:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
      additionalProperties: false
    ImportRowError:
      type: object
      properties:
        rowNumber:
          type: integer
          format: int32
        rawValues:
          type: object
          additionalProperties:
            type: string
            nullable: true
          nullable: true
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ImportFieldError'
          nullable: true
      additionalProperties: false
    ImportRowPreview:
      type: object
      properties:
        rowNumber:
          type: integer
          format: int32
        device:
          $ref: '#/components/schemas/ImportDevicePreview'
        brandId:
          type: string
          format: uuid
          nullable: true
        categoryId:
          type: string
          format: uuid
          nullable: true
        ownerId:
          type: string
          format: uuid
          nullable: true
        locationId:
          type: string
          format: uuid
          nullable: true
        networkId:
          type: string
          format: uuid
          nullable: true
      additionalProperties: false
    ImportStatus:
      enum:
        - Pending
        - Completed
        - PartialSuccess
        - Failed
      type: string
...
    MissingLookup:
      type: object
      properties:
        entityType:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
      additionalProperties: false
...
    PreviewImportResult:
      type: object
      properties:
        totalRows:
          type: integer
          format: int32
        validRows:
          type: array
          items:
            $ref: '#/components/schemas/ImportRowPreview'
          nullable: true
        invalidRows:
          type: array
          items:
            $ref: '#/components/schemas/ImportRowError'
          nullable: true
        lookupsToCreate:
          type: array
          items:
            $ref: '#/components/schemas/MissingLookup'
          nullable: true
      additionalProperties: false
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
    ValidationProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          nullable: true
      additionalProperties: { }
...